gmeet3/requirements.txt
click==8.0.1
datetime==4.3
pathlib==1.0.1
pyaudio==0.2.11
wave==0.0.2
sounddevice
langchain_community==0.3.24
langchain_core==0.3.66
opencv-python-headless
groq
flask
langchain_groq
sentence_transformers
faiss-cpu
werkzeug
pyaudio
click
opencv-python
Pillow
pydub
datetime
pathlib
selenium
wave
sounddevice
opencv-python-headless
pyannote.audio
requests
python-dotenv
onnxruntime
groq
flask
langchain_groq
langchain_core
sentence-transformers
faiss-cpu
numpy
hf_xet
